user_birthday.py
"""Opening a file. Reading the file. Appending the content of that file to new string. Printing that string. Taking input from user about his birthday. Checking if user's birthday in the file. If birthday is present in the file tell user about it, if birthday is not present tell user that too."""