Create area_test.py
Creating a class called rectangle and rectangle class takes two Attributes and these attributes are width and height. The rectangular class also has a method called area, This method returns the area of that rectangle.

Later we create another class using the unit test and that will test the rectangle class that we built initially.