Create palindrome.py
Here we are creating a class called string manipulator, This class also had a method called reverse and is palindrome, This reverse takes an attribute and reverse it and return the reversed string.

Is_palindrome Method checks if original string is equal to newly reversed string. 

Now the class test string manipulator is an instance of the unit test, This will test the string that we previously took so this will show the result if its passes or not.